requirements.txt
alabaster==0.7.10
Babel==2.4.0
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.3
django-jinja==2.3.1
django-modeltranslation==0.12.1
docutils==0.13.1
idna==2.5
imagesize==0.7.1
Jinja2==2.9.6
MarkupSafe==1.0
Pygments==2.2.0
PyMySQL==0.7.11
pytz==2017.2
requests==2.18.2
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.6.3
sphinx-rtd-theme==0.2.4
sphinxcontrib-websupport==1.0.1
urllib3==1.22
uWSGI==2.0.15
